+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Bug fixes
 
+- [Host] Show the host's memory usage instead of the sum of the VMs' memory usage (PR [#3924](https://github.com/vatesfr/xen-orchestra/pull/3924))
+
 ### Released packages
 
 - xo-server v5.35.0